BranchList component simplified and migrated into Branches component tree
